+++ src/config.go
@@ -1,0 +1,135 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/nbd-wtf/go-nostr"
+)
+
+type Config struct {
+	Relays     map[string]Relay   `json:"relays"`
+	Follows    map[string]Profile `json:"follows"`
+	PrivateKey string             `json:"privatekey"`
+	Updated    time.Time          `json:"updated"`
+	verbose    bool
+}
+
+func loadConfig(profile string) (*Config, error) {
+	dir, err := os.UserConfigDir()
+	if err != nil {
+		return nil, err
+	}
+	dir = filepath.Join(dir, "algia")
+
+	var fp string
+	if profile == "" {
+		fp = filepath.Join(dir, "config.json")
+	} else if profile == "?" {
+		names, err := filepath.Glob(filepath.Join(dir, "config-*.json"))
+		if err != nil {
+			return nil, err
+		}
+		for _, name := range names {
+			name = filepath.Base(name)
+			name = strings.TrimLeft(name[6:len(name)-5], "-")
+			fmt.Println(name)
+		}
+		os.Exit(0)
+	} else {
+		fp = filepath.Join(dir, "config-"+profile+".json")
+	}
+	os.MkdirAll(filepath.Dir(fp), 0700)
+
+	b, err := ioutil.ReadFile(fp)
+	if err != nil {
+		return nil, err
+	}
+	var cfg Config
+	err = json.Unmarshal(b, &cfg)
+	if err != nil {
+		return nil, err
+	}
+	return &cfg, nil
+}
+
+func (cfg *Config) FindRelay(r Relay) *nostr.Relay {
+	for k, v := range cfg.Relays {
+		if r.Write && !v.Write {
+			continue
+		}
+		if r.Search && !v.Search {
+			continue
+		}
+		if !r.Write && !v.Read {
+			continue
+		}
+		ctx := context.WithValue(context.Background(), "url", k)
+		relay, err := nostr.RelayConnect(ctx, k)
+		if err != nil {
+			if cfg.verbose {
+				fmt.Fprintln(os.Stderr, err.Error())
+			}
+			continue
+		}
+		return relay
+	}
+	return nil
+}
+
+func (cfg *Config) Do(r Relay, f func(*nostr.Relay)) {
+	var wg sync.WaitGroup
+	for k, v := range cfg.Relays {
+		if r.Write && !v.Write {
+			continue
+		}
+		if r.Search && !v.Search {
+			continue
+		}
+		if !r.Write && !v.Read {
+			continue
+		}
+		wg.Add(1)
+		go func(wg *sync.WaitGroup, k string, v Relay) {
+			defer wg.Done()
+			ctx := context.WithValue(context.Background(), "url", k)
+			relay, err := nostr.RelayConnect(ctx, k)
+			if err != nil {
+				if cfg.verbose {
+					fmt.Fprintln(os.Stderr, err)
+				}
+				return
+			}
+			f(relay)
+			relay.Close()
+		}(&wg, k, v)
+	}
+	wg.Wait()
+}
+
+func (cfg *Config) save(profile string) error {
+	dir, err := os.UserConfigDir()
+	if err != nil {
+		return err
+	}
+	dir = filepath.Join(dir, "algia")
+
+	var fp string
+	if profile == "" {
+		fp = filepath.Join(dir, "config.json")
+	} else {
+		fp = filepath.Join(dir, "config-"+profile+".json")
+	}
+	b, err := json.MarshalIndent(&cfg, "", "  ")
+	if err != nil {
+		return err
+	}
+	return ioutil.WriteFile(fp, b, 0644)
+}
