+++ src/main.go
@@ -1,0 +1,127 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/urfave/cli/v2"
+)
+
+const name = "algia"
+
+const version = "0.0.1"
+
+var revision = "HEAD"
+
+func main() {
+	app := &cli.App{
+		Description: "A cli application for nostr",
+		Flags: []cli.Flag{
+			&cli.StringFlag{Name: "a", Usage: "profile name"},
+			&cli.BoolFlag{Name: "V", Usage: "verbose"},
+		},
+		Commands: []*cli.Command{
+			{
+				Name:    "timeline",
+				Aliases: []string{"tl"},
+				Usage:   "show timeline",
+				Flags: []cli.Flag{
+					&cli.IntFlag{Name: "n", Value: 30, Usage: "number of items"},
+					&cli.BoolFlag{Name: "json", Usage: "output JSON"},
+					&cli.BoolFlag{Name: "extra", Usage: "extra JSON"},
+				},
+				Action: doTimeline,
+			},
+			{
+				Name:    "post",
+				Aliases: []string{"n"},
+				Flags: []cli.Flag{
+					&cli.BoolFlag{Name: "stdin"},
+					&cli.StringFlag{Name: "sensitive"},
+				},
+				Usage:     "post new note",
+				UsageText: "algia post [note text]",
+				HelpName:  "post",
+				ArgsUsage: "[note text]",
+				Action:    doPost,
+			},
+			{
+				Name:    "reply",
+				Aliases: []string{"r"},
+				Flags: []cli.Flag{
+					&cli.BoolFlag{Name: "stdin"},
+					&cli.StringFlag{Name: "id", Required: true},
+					&cli.BoolFlag{Name: "quote"},
+				},
+				Usage:     "reply to the note",
+				UsageText: "algia reply --id [id] [note text]",
+				HelpName:  "reply",
+				ArgsUsage: "[note text]",
+				Action:    doReply,
+			},
+			{
+				Name:    "repost",
+				Aliases: []string{"b"},
+				Flags: []cli.Flag{
+					&cli.StringFlag{Name: "id", Required: true},
+				},
+				Usage:     "repost the note",
+				UsageText: "algia repost --id [id]",
+				HelpName:  "repost",
+				Action:    doRepost,
+			},
+			{
+				Name:    "like",
+				Aliases: []string{"l"},
+				Flags: []cli.Flag{
+					&cli.StringFlag{Name: "id", Required: true},
+				},
+				Usage:     "like the note",
+				UsageText: "algia like --id [id]",
+				HelpName:  "lite",
+				Action:    doLike,
+			},
+			{
+				Name:    "delete",
+				Aliases: []string{"d"},
+				Flags: []cli.Flag{
+					&cli.StringFlag{Name: "id", Required: true},
+				},
+				Usage:     "delete the note",
+				UsageText: "algia delete --id [id]",
+				HelpName:  "delete",
+				Action:    doDelete,
+			},
+			{
+				Name:    "search",
+				Aliases: []string{"s"},
+				Flags: []cli.Flag{
+					&cli.IntFlag{Name: "n", Value: 30, Usage: "number of items"},
+					&cli.BoolFlag{Name: "json", Usage: "output JSON"},
+					&cli.BoolFlag{Name: "extra", Usage: "extra JSON"},
+				},
+				Usage:     "search notes",
+				UsageText: "algia search [words]",
+				HelpName:  "search",
+				Action:    doSearch,
+			},
+		},
+		Before: func(cCtx *cli.Context) error {
+			profile := cCtx.String("a")
+			cfg, err := loadConfig(profile)
+			if err != nil {
+				return err
+			}
+			cCtx.App.Metadata = map[string]any{
+				"config": cfg,
+			}
+			cfg.verbose = cCtx.Bool("V")
+			return nil
+		},
+	}
+
+	if err := app.Run(os.Args); err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}
