+++ src/post.go
@@ -1,0 +1,75 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"strings"
+	"sync/atomic"
+	"time"
+
+	"github.com/nbd-wtf/go-nostr"
+	"github.com/nbd-wtf/go-nostr/nip19"
+	"github.com/urfave/cli/v2"
+)
+
+func doPost(cCtx *cli.Context) error {
+	stdin := cCtx.Bool("stdin")
+	if !stdin && cCtx.Args().Len() == 0 {
+		return cli.ShowSubcommandHelp(cCtx)
+	}
+	sensitive := cCtx.String("sensitive")
+
+	cfg := cCtx.App.Metadata["config"].(*Config)
+
+	var sk string
+	if _, s, err := nip19.Decode(cfg.PrivateKey); err != nil {
+		return err
+	} else {
+		sk = s.(string)
+	}
+	ev := nostr.Event{}
+	if pub, err := nostr.GetPublicKey(sk); err == nil {
+		if _, err := nip19.EncodePublicKey(pub); err != nil {
+			return err
+		}
+		ev.PubKey = pub
+	} else {
+		return err
+	}
+
+	if stdin {
+		b, err := ioutil.ReadAll(os.Stdin)
+		if err != nil {
+			return err
+		}
+		ev.Content = string(b)
+	} else {
+		ev.Content = strings.Join(cCtx.Args().Slice(), "\n")
+	}
+
+	if sensitive != "" {
+		ev.Tags = ev.Tags.AppendUnique(nostr.Tag{"content-warning", sensitive})
+	}
+
+	ev.CreatedAt = time.Now()
+	ev.Kind = nostr.KindTextNote
+	ev.Sign(sk)
+
+	var success atomic.Int64
+	cfg.Do(Relay{Write: true}, func(relay *nostr.Relay) {
+		status := relay.Publish(context.Background(), ev)
+		if cfg.verbose {
+			fmt.Fprintln(os.Stderr, relay.URL, status)
+		}
+		if status != nostr.PublishStatusFailed {
+			success.Add(1)
+		}
+	})
+	if success.Load() == 0 {
+		return errors.New("cannot post")
+	}
+	return nil
+}
