+++ src/profile.go
@@ -1,0 +1,11 @@
+package main
+
+type Profile struct {
+	Website     string `json:"website"`
+	Nip05       string `json:"nip05"`
+	Picture     string `json:"picture"`
+	Lud16       string `json:"lud16"`
+	DisplayName string `json:"display_name"`
+	About       string `json:"about"`
+	Name        string `json:"name"`
+}
