+++ src/relay.go
@@ -1,0 +1,7 @@
+package main
+
+type Relay struct {
+	Read   bool `json:"read"`
+	Write  bool `json:"write"`
+	Search bool `json:"search"`
+}
