+++ src/repost.go
@@ -1,0 +1,75 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"sync/atomic"
+	"time"
+
+	"github.com/nbd-wtf/go-nostr"
+	"github.com/nbd-wtf/go-nostr/nip19"
+	"github.com/urfave/cli/v2"
+)
+
+func doRepost(cCtx *cli.Context) error {
+	id := cCtx.String("id")
+
+	cfg := cCtx.App.Metadata["config"].(*Config)
+
+	ev := nostr.Event{}
+	var sk string
+	if _, s, err := nip19.Decode(cfg.PrivateKey); err != nil {
+		return err
+	} else {
+		sk = s.(string)
+	}
+	if pub, err := nostr.GetPublicKey(sk); err == nil {
+		if _, err := nip19.EncodePublicKey(pub); err != nil {
+			return err
+		}
+		ev.PubKey = pub
+	} else {
+		return err
+	}
+
+	if _, tmp, err := nip19.Decode(id); err == nil {
+		ev.Tags = ev.Tags.AppendUnique(nostr.Tag{"e", tmp.(string)})
+	} else {
+		ev.Tags = ev.Tags.AppendUnique(nostr.Tag{"e", id})
+	}
+	filter := nostr.Filter{
+		Kinds: []int{nostr.KindTextNote},
+		IDs:   []string{id},
+	}
+
+	ev.CreatedAt = time.Now()
+	ev.Kind = nostr.KindBoost
+	ev.Content = ""
+
+	var first atomic.Bool
+	first.Store(true)
+
+	var success atomic.Int64
+	cfg.Do(Relay{Write: true}, func(relay *nostr.Relay) {
+		if first.Load() {
+			for _, tmp := range relay.QuerySync(context.Background(), filter) {
+				ev.Tags = ev.Tags.AppendUnique(nostr.Tag{"p", tmp.ID})
+			}
+			first.Store(false)
+			ev.Sign(sk)
+		}
+		status := relay.Publish(context.Background(), ev)
+		if cfg.verbose {
+			fmt.Fprintln(os.Stderr, relay.URL, status)
+		}
+		if status != nostr.PublishStatusFailed {
+			success.Add(1)
+		}
+	})
+	if success.Load() == 0 {
+		return errors.New("cannot repost")
+	}
+	return nil
+}
